+++ Net_Core/Web.Angular/projects/gw-frontend/src/app/routes/testing-routes.ts
@@ -2,7 +2,9 @@ import { Routes } from '@angular/router';
 // Library
 import { AuthGuard } from '@growthware/common/services';
 import { TestLoggingComponent } from '@growthware/core/testing';
+import { TestModalComponent } from '@growthware/core/testing';
 
 export const testingRoutes: Routes = [
-    { path: 'logging', component: TestLoggingComponent, canActivate: [AuthGuard] }
+    { path: 'logging', component: TestLoggingComponent, canActivate: [AuthGuard] },
+    { path: 'modal', component: TestModalComponent, canActivate: [AuthGuard] }
 ];