+++ Net_Core/Web.Angular/projects/gw-lib/core/testing/public-api.ts
@@ -3,3 +3,4 @@
  */
 
 export * from './src/c/test-logging/test-logging.component';
+export * from './src/c/test-modal/test-modal.component';