+++ Net_Core/Web.Angular/projects/gw-lib/core/testing/src/c/test-modal/test-modal.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'gw-core-test-modal',
+  standalone: true,
+  imports: [],
+  templateUrl: './test-modal.component.html',
+  styleUrl: './test-modal.component.scss'
+})
+export class TestModalComponent {
+
+}